Various light refactoring
