+++ src/rv2/core/annotation_source.py
@@ -5,10 +5,7 @@ class AnnotationSource(ABC):
     """A source of annotations for a project.
 
     An AnnotationSource is a read/write source of annotations for a project
-    that could be backed by a file, a database, an API, etc. An example of an
-    annotation is a bounding box in the case of object detection or a raster in
-    case of segmentation. It is something that is predicted by a model or
-    provided by human annotators for the sake of training.
+    that could be backed by a file, a database, an API, etc.
     """
 
     @abstractmethod
@@ -42,7 +39,6 @@ class AnnotationSource(ABC):
         """
         pass
 
-    # TODO just absorb this into save?
     @abstractmethod
     def post_process(self):
         """Perform some preprocessing operation before saving."""
